+++ skyperious/live.py
@@ -8,7 +8,7 @@ Released under the MIT License.
 
 @author      Erki Suurjaak
 @created     08.07.2020
-@modified    14.03.2021
+@modified    01.08.2021
 ------------------------------------------------------------------------------
 """
 import base64
@@ -41,6 +41,9 @@ from . import skypedata
 
 
 logger = logging.getLogger(__name__)
+
+# Avoid BeautifulSoup popup warnings like MarkupResemblesLocatorWarning
+warnings.filterwarnings("ignore", category=UserWarning, module="bs4")
 
 
 class SkypeLogin(object):
@@ -247,18 +250,10 @@ class SkypeLogin(object):
         dbitem = self.convert(table, item, parent=parent)
         if dbitem is None: return self.SAVE.SKIP
         result, dbitem1, table = None, dict(dbitem), table.lower()
-        do_fix_bot = False
 
         identity = dbitem["skypename"] if table in ("accounts", "contacts") else \
                    dbitem["identity"]  if "chats" == table else dbitem.get("pk_id")
         dbitem0 = self.cache[table].get(identity)
-
-        if table in ("chats", "contacts") and not dbitem0 \
-        and (identity or "").startswith(skypedata.ID_PREFIX_BOT):
-            # Handle bot accounts synced before v4.5.1
-            dbitem0 = self.cache[table].get(identity[len(skypedata.ID_PREFIX_BOT):])
-            if dbitem0: self.cache[table].pop(dbitem0.get("identity"), None)
-            do_fix_bot = dbitem0 and "contacts" == table
 
         if "contacts" == table and not dbitem0 \
         and not (identity or "").startswith(skypedata.ID_PREFIX_BOT):
@@ -311,7 +306,10 @@ class SkypeLogin(object):
         if dbitem0:
             dbitem["id"] = dbitem1["id"] = dbitem0["id"]
             if "messages" == table and dbitem0.get("remote_id") == dbitem.get("remote_id") \
-            and (dbitem0["pk_id"] != dbitem["pk_id"] or dbitem0["body_xml"] != dbitem["body_xml"]):
+            and (dbitem0["pk_id"] != dbitem["pk_id"] or dbitem0["body_xml"] != dbitem["body_xml"]) \
+            and not (isinstance(item, skpy.SkypeFileMsg) # body_xml contains url, starting from v4.9
+                     and dbitem0["pk_id"] == dbitem["pk_id"]
+                     and dbitem0["timestamp__ms"] == dbitem["timestamp__ms"]):
                 # Different messages with same remote_id -> edited or deleted message
                 dbitem["edited_by"] = dbitem["author"]
                 dbitem["edited_timestamp"] = max(dbitem["timestamp"], dbitem0["timestamp"])
@@ -348,9 +346,6 @@ class SkypeLogin(object):
                 self.cache["contacts"][identity] = cacheitem # For name lookup
         if "messages" == table:
             self.msg_lookups[dbitem["timestamp__ms"]].append(cacheitem)
-
-        if do_fix_bot: fix_bot_identity(self.db, identity)
-
         if "chats" == table:
             self.insert_participants(item, dbitem, dbitem0)
             self.insert_chats(item, dbitem, dbitem0)
@@ -425,13 +420,7 @@ class SkypeLogin(object):
                 existing.update(x["identity"] for x in cursor)
 
             for p in participants:
-                if p["identity"] not in existing \
-                and p["identity"].startswith(skypedata.ID_PREFIX_BOT) \
-                and p["identity"][len(skypedata.ID_PREFIX_BOT):] in existing:
-                    # Handle bot accounts synced before v4.5.1
-                    p0 = existing.get(p["identity"][len(skypedata.ID_PREFIX_BOT):])
-                    self.db.update_row("participants", p, p0, log=False)
-                elif p["identity"] not in existing:
+                if p["identity"] not in existing:
                     self.db.insert_row("participants", p, log=False)
                 if p["identity"] != self.skype.userId \
                 and (p["identity"] not in self.cache["contacts"]
@@ -523,7 +512,7 @@ class SkypeLogin(object):
                 result.update(emails=" ".join(item.raw["emails"]))
 
             if item.avatar: # https://avatar.skype.com/v1/avatars/username/public
-                raw = self.download_media(item.avatar)
+                raw = self.download_content(item.avatar)
                 # main.db has NULL-byte in front of image binary
                 if raw: result.update(avatar_image="\0" + raw)
 
@@ -559,7 +548,7 @@ class SkypeLogin(object):
                     if creator: result.update(creator=creator)
                 if item.picture:
                     # https://api.asm.skype.com/v1/objects/0-weu-d14-abcdef..
-                    raw = self.get_api_media(item.picture, category="avatar")
+                    raw = self.get_api_content(item.picture, category="avatar")
                     # main.db has NULL-byte in front of image binary
                     if raw: result.update(meta_picture="\0" + raw)
             else: result = None
@@ -639,9 +628,11 @@ class SkypeLogin(object):
                 # SkypeFileMsg(id='1594466346559', type='RichText/Media_GenericFile', time=datetime.datetime(2020, 7, 11, 11, 18, 19, 39000), clientId='8167024171841589273', userId='username', chatId='8:username', content='<URIObject uri="https://api.asm.skype.com/v1/objects/0-weu-d3-abcdef.." url_thumbnail="https://api.asm.skype.com/v1/objects/0-weu-d3-abcdef../views/original" type="File.1" doc_id="0-weu-d3-abcdef..">To view this file, go to: <a href="https://login.skype.com/login/sso?go=webclient.xmm&amp;docid=0-weu-d3-abcdef..">https://login.skype.com/login/sso?go=webclient.xmm&amp;docid=0-weu-d3-abcdef..</a><OriginalName v="f.txt"></OriginalName><FileSize v="447"></FileSize></URIObject>', file=File(name='f.txt', size='447', urlFull='https://api.asm.skype.com/v1/objects/0-weu-d3-abcdef..', urlThumb='https://api.asm.skype.com/v1/objects/0-weu-d3-abcdef../views/original', urlView='https://login.skype.com/login/sso?go=webclient.xmm&docid=0-weu-d3-abcdef..'))
 
                 filename, filesize = (item.file.name, item.file.size) if item.file else (None, None)
+                fileurl = item.file.urlFull
                 result.update(chatmsg_type=skypedata.CHATMSG_TYPE_SPECIAL, type=skypedata.MESSAGE_TYPE_FILE,
-                              body_xml='<files><file index="0" size="%s">%s</file></files>' % 
+                              body_xml='<files><file index="0" size="%s" url="%s">%s</file></files>' % 
                                        (urllib.quote(util.to_unicode(filesize or 0)),
+                                        urllib.quote(util.to_unicode(fileurl or ""), ":/"),
                                         util.to_unicode(filename or "file").replace("<", "&lt;").replace(">", "&gt;")))
 
             elif isinstance(item, skpy.msg.SkypeTopicPropertyMsg):
@@ -892,44 +883,40 @@ class SkypeLogin(object):
 
 
 
-    def get_api_media(self, url, category=None):
+    def get_api_content(self, url, category=None):
         """
-        Returns media raw binary from Skype API URL via login, or None.
+        Returns content raw binary from Skype API URL via login, or None.
 
-        @param   category  type of media, e.g. "avatar" for avatar image
+        @param   category  type of content, e.g. "avatar" for avatar image,
+                           "file" for shared file
         """
-        url = make_media_url(url, category)
+        url = make_content_url(url, category)
         urls = [url]
         # Some images appear to be available on one domain, some on another
         if not url.startswith("https://experimental-api.asm"):
             url0 = re.sub(r"https\:\/\/.+\.asm", "https://experimental-api.asm", url)
             urls.insert(0, url0)
         for url in urls:
-            raw = self.download_media(url)
+            raw = self.download_content(url, category)
             if raw: return raw
 
 
-    def download_media(self, url):
+    def download_content(self, url, category=None):
         """Downloads and returns media raw binary from Skype URL via login, or None."""
         try:
             r = self.request(self.skype.conn, "GET", url,
                              auth=skpy.SkypeConnection.Auth.Authorize,
                              __retry=False)
             hdr = lambda r: r.headers.get("content-type", "")
-            if r.ok and r.content and any(x in hdr(r) for x in ("image", "audio", "video")):
+            if r.ok and r.content \
+            and ("file" == category or any(x in hdr(r) for x in ("image", "audio", "video"))):
                 return r.content
         except Exception: pass
 
 
     def get_contact(self, identity):
         """Returns cached contact, if any."""
-        if not identity: return None
-        result = self.cache["contacts"].get(identity)
-        if not result: # Handle bot accounts synced before v4.5.1
-            result = self.cache["contacts"].get(skypedata.ID_PREFIX_BOT + identity)
-        if not result and identity and identity.startswith(skypedata.ID_PREFIX_BOT):
-            result = self.cache["contacts"].get(identity[len(skypedata.ID_PREFIX_BOT):])
-        return result
+        return self.cache["contacts"].get(identity) if identity else None
 
 
     def get_contact_name(self, identity):
@@ -1097,7 +1084,7 @@ class SkypeExport(skypedata.SkypeDatabase):
                     else:
                         try:
                             chat = self.export_finalize_chat(chat)
-                            self.update_row("conversations", chat, chat, log=False)
+                            self.update_row("conversations", chat, {"id": chat["id"]}, log=False)
                         except Exception:
                             logger.warn("Error updating chat %s.", chat, exc_info=True)
                     edited_msgs.clear()
@@ -1560,64 +1547,13 @@ def process_message_edit(msg):
         if BeautifulSoup: logger.warn("Error parsing edited timestamp from %s.", msg, exc_info=True)
 
 
-def fix_bot_identity(db, *identities):
+def make_content_url(url, category=None):
     """
-    Replaces bot contact identity with "28:"-prefix in all possible tables,
-    for data synced before v4.5.1.
-
-    @param   db          SkypeDatabase instance
-    @param   identities  one or more bot identities, with or without bot-prefix
-    """
-    SIMPLES = {"alerts":             ["partner_name"],
-               "callmembers":        ["identity", "real_identity"],
-               "calls":              ["host_identity"],
-               "chats":              ["dialog_partner"],
-               "contacts":           ["skypename"],
-               "messageannotations": ["author"],
-               "messages":           ["author", "dialog_partner", "edited_by"],
-               "participants":       ["identity"],
-               "transfers":          ["partner_handle"], }
-    SPACEDS = {"chats":              ["posters", "participants", "activemembers"],
-               "messages":           ["identities"], }
-    changeds = set()
-    for identity2 in filter(bool, identities):
-        if not identity2.startswith(skypedata.ID_PREFIX_BOT):
-            identity1, identity2 = identity2, skypedata.ID_PREFIX_BOT + identity2
-        else: identity1 = identity2[len(skypedata.ID_PREFIX_BOT):]
-        for table, cols in SIMPLES.items():
-            if table not in db.tables: continue # for table, cols
-            for col in cols:
-                sql = "UPDATE %s SET %s = ? WHERE %s = ?" % (table, col, col)
-                args = (identity2, identity1)
-                if db.execute(sql, args).rowcount: changeds.add(table)
-        for table, cols in SPACEDS.items():
-            if table not in db.tables: continue # for table, cols
-            for col in cols:
-                args = ("%" + identity1 + "%", )
-                sql = "SELECT id, %s FROM %s WHERE %s LIKE ?" % (col, table, col)
-                for row in db.execute(sql, args).fetchall():
-                    vals = row[col].split()
-                    if identity1 in vals:
-                        v2 = " ".join(identity2 if v == identity1 else v for v in vals)
-                        args = (v2, row["id"])
-                        sql = "UPDATE %s SET %s = ? WHERE id = ?" % (table, col)
-                        db.execute(sql, args)
-                        changeds.add(table)
-        args = (skypedata.CONTACT_TYPE_BOT, identity2)
-        if db.execute("UPDATE contacts SET type = ? WHERE skypename = ?", args).rowcount:
-            changeds.add("contacts")
-    if changeds: changeds.add("conversations")
-    for table in changeds:
-        for d in (db.table_rows, db.table_objects): d.pop(table, None)
-
-
-def make_media_url(url, category=None):
-    """
-    Returns URL with appropriate path appended if Skype shared media URL,
+    Returns URL with appropriate path appended if Skype shared content URL,
     e.g. "https://api.asm.skype.com/v1/objects/0-weu-d11-../views/imgpsh_fullsize"
     for  "https://api.asm.skype.com/v1/objects/0-weu-d11-..".
 
-    @param   category  type of media, e.g. "avatar" for avatar image
+    @param   category  type of content, e.g. "avatar" for avatar image
     """
     if not url or "api.asm.skype.com/" not in url: return url
 
@@ -1629,6 +1565,8 @@ def make_media_url(url, category=None):
         url += "/views/video"
     elif "sticker" == category and not url.endswith("/views/thumbnail"):
         url += "/views/thumbnail"
+    elif "file"    == category and not url.endswith("/views/original"):
+        url += "/views/original"
     elif "/views/" not in url:
         url += "/views/imgpsh_fullsize"